+++ litebox_broker_transport/src/lib.rs
@@ -14,6 +14,3 @@ pub mod shared_memory;
 
 #[cfg(all(feature = "unix", unix))]
 pub mod unix_socket;
-
-#[cfg(all(feature = "unix-shared-buffer", target_os = "linux"))]
-pub mod unix_shared_buffer;
